pyproject.toml
[tool.poetry]
name = "izba-docs-api"
version = "0.1.0"
description = ""
authors = ["Przemek Kaliś <przemek.kalis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
django-cors-headers = "^3.7.0"
django-environ = "^0.4.5"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.1"
sentry-sdk = "^1.3.1"

[tool.poetry.dev-dependencies]
bandit = "^1.7.0"
black = "^21.7b0"
isort = "^5.9.2"
pre-commit = "^2.13.0"
pylint = "^2.9.6"
pylint-django = "^2.4.4"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-django = "^4.4.0"
schema = "^0.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
